requirements.txt
feedparser
python-dotenv
google-generativeai
